% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_calls_saturn.R
\name{p1x}
\alias{p1x}
\title{P1X Generic Call}
\usage{
p1x(net1, net2, mat1, mat2, key)
}
\arguments{
\item{net1}{ufs of the main network}

\item{net2}{ufs of the second network}

\item{mat1}{ufm of the first matrix (if not detected automatically)}

\item{mat2}{ufm of the first matrix (if not detected automatically)}

\item{key}{key file to use}
}
\value{
nothing executes call
}
\description{
P1X Generic Call
}
