% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_coordinates.r
\name{prepare_coordinates}
\alias{prepare_coordinates}
\title{Prepare coordinates file}
\usage{
prepare_coordinates(dat, wd = ".", save_file = TRUE, output_file = if
  (save_file) "Coordinates.csv")
}
\description{
This function allows you to read coordinates.dat file and
extract the coordinates of all the nodes in a saturn network
}
\examples{
prepare_coordinates("2014_coordinates.dat")
}
