% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufm.R
\name{read_ufm}
\alias{read_ufm}
\title{Read UFM function}
\usage{
read_ufm(file, remove_txt = TRUE, clean_up = TRUE)
}
\arguments{
\item{file}{The file to read.}

\item{remove_txt}{when set to TRUE it will remove the interim txt files}

\item{clean_up}{when TRUE removes the VDU, KEY and LPX files}
}
\description{
This function allows you to read UFM files into R.
}
\keyword{read,}
\keyword{ufm}
