% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ufs.R
\name{read_ufs}
\alias{read_ufs}
\title{Read UFS function}
\usage{
read_ufs(file, clean_up = TRUE, load_geometry = FALSE,
  selection_mode = "all_links", step = 9)
}
\arguments{
\item{file}{The file to read.}

\item{clean_up}{when TRUE removes the VDU, KEY and LPX files}

\item{load_geometry}{it uses satdb to load the geometry and converts the object into a sf}

\item{selection_mode}{Define how to query the ufs with the options 'all_links'(default), 'simulation_links'
'simulation_turns' or 'centroid_connectors'.}

\item{step}{how many records to read by iteration, max is 9, use 1 for stability.}
}
\description{
This function allows you to read UFS assignment data into R.
}
\keyword{read,}
\keyword{ufm}
