% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn}
\alias{saturn}
\title{saturn function}
\usage{
saturn(DAT = "", UFM = "", PASSQ = NULL,
  PATH = "C:/SATWIN/XEXES_11.3.12W_MC", remove_bat = TRUE)
}
\arguments{
\item{DAT}{The network file.}

\item{UFM}{The Matrix file.}

\item{PASSQ}{The PASSQ network file (optional).}

\item{PATH}{The path to the XEXES version of preference (Default: "C:/SATWIN/XEXES_11.3.12W_MC").}

\item{remove_bat}{when set to TRUE it will remove the .bat file generated.}
}
\description{
This runs saturn assignements using the SATURN.bat file stored in the XEXES folder.
}
\examples{
saturn('./Data/AM-reassignment/2014_Base_AM.dat','./Data/AM-reassignment/2014_BaseCosts_AM.ufm','./Data/AM-reassignment/2014_Base_AM_PassQ.dat')
}
\keyword{assign}
\keyword{saturn,}
