% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{saturnR_example}
\alias{saturnR_example}
\title{Get path to saturnR example}
\usage{
saturnR_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
saturnR comes bundled with some example files in its `inst/extdata`
directory. This function make them easy to access. (Obtained from readxl package)
}
\examples{
readxl_example()
}
