% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufs_to_sf.R
\name{ufs_as_sf}
\alias{ufs_as_sf}
\title{ufs_to_sf}
\usage{
ufs_as_sf(file, ...)
}
\arguments{
\item{file}{The file to read.}

\item{...}{Arguments to be passed to get-coordinates}
}
\description{
This function allows you to read UFS/N file's into an R SF.
}
\examples{
ufs_to_sf("file.UFS")
}
\keyword{coordinates,}
\keyword{read,}
\keyword{sf}
\keyword{ufs,}
