% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufm.R
\name{write_ufm}
\alias{write_ufm}
\title{Write UFM function}
\usage{
write_ufm(x, file, remove_txt = TRUE, clean_up = TRUE, stack = FALSE)
}
\arguments{
\item{x}{A data,table object with three columns O,D,Trips.}

\item{file}{The file to write, extension mandatory}

\item{remove_txt}{when set to TRUE it will remove the interim txt files}

\item{clean_up}{when TRUE removes the VDU, KEY and LPX files}

\item{stack}{wether or not to stack the function. Changes to true for 4 column inputs}
}
\description{
This function allows you to write data frames into UFM files.
}
\keyword{ufm}
\keyword{write,}
