% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufs_to_sf.R
\name{get_coordinates}
\alias{get_coordinates}
\title{get coordinates}
\usage{
get_coordinates(file, remove_txt = TRUE, clean_up = TRUE)
}
\arguments{
\item{file}{The file to read.}

\item{remove_txt}{when TRUE removes the interim txt}

\item{clean_up}{when TRUE removes the VDU, KEY and LPX files}
}
\description{
This function allows you to read UFS/N file's into a df.
}
\keyword{coordinates}
\keyword{read,}
\keyword{ufs,}
