% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_coordinates.r
\name{prepare_coordinates}
\alias{prepare_coordinates}
\title{Prepare coordinates file}
\usage{
prepare_coordinates(dat, wd = ".", save_file = TRUE, output_file = if
  (save_file) "Coordinates.csv")
}
\arguments{
\item{dat}{character; the name of the file to read e.g. "coordinates.dat"}

\item{wd}{character; the working directory where the coordinates file exist}

\item{save_file}{logical; TRUE to save the final file as .csv, else FALSE; default to TRUE}

\item{output_file}{the name of the output csv file}
}
\description{
This function allows you to read coordinates.dat file and
extract the coordinates of all the nodes in a saturn network
Use ufs_to_sf for link coordinates and an SF object
}
\examples{
prepare_coordinates("2014_coordinates.dat")
}
\keyword{coordinates,}
\keyword{nodes}
