% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ufm.R
\name{read_ufm}
\alias{read_ufm}
\title{Read UFM function}
\usage{
read_ufm(file, MX = "C:\\\\SATWIN\\\\XEXES_11.3.12W_MC\\\\$MX.exe",
  remove_txt = TRUE, clean_up = TRUE)
}
\arguments{
\item{file}{The file to read.}

\item{MX}{Path to the MX exe folder}

\item{remove_txt}{when set to TRUE it will remove the interim txt files}

\item{clean_up}{when TRUE removes the VDU, KEY and LPX files}
}
\description{
This function allows you to express your love of cats.
}
\examples{
read_ufm("file.UFM")
}
\keyword{read,}
\keyword{ufm}
