% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ufs.R
\name{read_ufs}
\alias{read_ufs}
\title{Read UFS function}
\usage{
read_ufs(file, P1X = "C:\\\\SATWIN\\\\XEXES_11.3.12W_MC\\\\$P1X.exe",
  remove_txt = TRUE, clean_up = TRUE)
}
\arguments{
\item{file}{The file to read.}

\item{remove_txt}{when set to TRUE it will remove the interim txt files}

\item{clean_up}{when TRUE removes the VDU, KEY and LPX files}

\item{MX}{Path to the MX exe folder}
}
\description{
This function allows you to read UFS assignment data into R.
}
\examples{
read_ufs("file.UFM")

}
\keyword{read,}
\keyword{ufm}
