% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_dist.R
\name{skim_dist}
\alias{skim_dist}
\title{Skim Distance UFM function}
\usage{
skim_dist(file,
  SATLOOK = "C:\\\\SATWIN\\\\XEXES_11.3.12W_MC\\\\$SATLOOK.exe",
  remove_ufm = TRUE, clean_up = TRUE)
}
\arguments{
\item{file}{The file to read.}

\item{remove_ufm}{when set to TRUE it will remove the interim ufm files}

\item{clean_up}{when TRUE removes the VDU, KEY and LPL files}

\item{SKIMDIST}{Path to the SKIMDIST.bat folder}
}
\description{
This function allows you to skim distances (km) from UFS files into R. Note it needs UFC/O
}
\examples{
skim_dist("file.UFS")
}
\keyword{distance}
\keyword{read,}
\keyword{ufs,}
