% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ufm.R
\name{write_ufm}
\alias{write_ufm}
\title{Read UFM function}
\usage{
write_ufm(x, file, MX = "C:\\\\SATWIN\\\\XEXES_11.3.12W_MC\\\\$MX.exe",
  remove_txt = TRUE, clean_up = TRUE)
}
\arguments{
\item{x}{A data,table object with three columns O,D,Trips.}

\item{file}{The file to write, extension mandatory.}

\item{MX}{Path to the MX exe folder}

\item{remove_txt}{when set to TRUE it will remove the interim txt files}

\item{clean_up}{when TRUE removes the VDU, KEY and LPX files}
}
\description{
This function allows you to write data frames into UFM files.
}
\examples{
read_ufm("file.UFM")
}
\keyword{ufm}
\keyword{write,}
